feat(paprika): user params crud. menu calculation based on user params. user calculation saving.

+++ src/main/kotlin/com/paprika/database/models/user/UserEatingsParamsModel.kt
@@ -5,7 +5,7 @@ import com.paprika.utils.database.BaseIntIdTable
 object UserEatingsParamsModel: BaseIntIdTable() {
     val user = reference("user", UserModel)
     val name = text("name")
-    val size = integer("size")
+    val size = double("size")
     val type = text("type")
     val difficulty = integer("difficulty").default(0)
     val dishCount = integer("dish_count").default(0)
